// src/types/rollup.d.ts
export interface SourceMap {
  file: string;
  mappings: string;
  names: string[];
  sources: string[];
  sourcesContent: string[];
  version: number;
  toString(): string;
  toUrl(): string;
}
// ...
export interface ModuleJSON extends TransformModuleJSON {
  alwaysRemovedCode: [number, number][];
  ast: AcornNode;
  dependencies: string[];
  id: string;
  transformFiles: EmittedFile[] | undefined;
}
// ...
export interface OutputPlugin
  extends Partial<OutputPluginHooks>,
    Partial<OutputPluginValueHooks> {
  name: string;
}
// ...
export type GlobalsOption =
  | { [name: string]: string }
  | ((name: string) => string);
export type InputOption = string | string[] | { [entryAlias: string]: string };
export type ManualChunksOption =
  | { [chunkAlias: string]: string[] }
  | GetManualChunk;
export type ModuleSideEffectsOption =
  | boolean
  | "no-external"
  | string[]
  | HasModuleSideEffects;
export type PreserveEntrySignaturesOption =
  | false
  | "strict"
  | "allow-extension";
export type SourcemapPathTransformOption = (
  relativeSourcePath: string,
  sourcemapPath: string
) => string;
// ...
export type ModuleFormat =
  | InternalModuleFormat
  | "commonjs"
  | "esm"
  | "module"
  | "systemjs";

export type OptionsPaths = Record<string, string> | ((id: string) => string);

export type InteropType =
  | boolean
  | "auto"
  | "esModule"
  | "default"
  | "defaultOnly";

export type GetInterop = (id: string | null) => InteropType;

export interface OutputOptions {
  amd?: {
    define?: string;
    id?: string;
  };
  assetFileNames?: string | ((chunkInfo: PreRenderedAsset) => string);
  banner?: string | (() => string | Promise<string>);
  chunkFileNames?: string | ((chunkInfo: PreRenderedChunk) => string);
  compact?: boolean;
  // only required for bundle.write
  dir?: string;
  /** @deprecated Use the "renderDynamicImport" plugin hook instead. */
  dynamicImportFunction?: string;
  entryFileNames?: string | ((chunkInfo: PreRenderedChunk) => string);
  esModule?: boolean;
  exports?: "default" | "named" | "none" | "auto";
  extend?: boolean;
  externalLiveBindings?: boolean;
  // only required for bundle.write
  file?: string;
  footer?: string | (() => string | Promise<string>);
  format?: ModuleFormat;
  freeze?: boolean;
  globals?: GlobalsOption;
  hoistTransitiveImports?: boolean;
  indent?: string | boolean;
  inlineDynamicImports?: boolean;
  interop?: InteropType | GetInterop;
  intro?: string | (() => string | Promise<string>);
  manualChunks?: ManualChunksOption;
  minifyInternalExports?: boolean;
  name?: string;
  namespaceToStringTag?: boolean;
  noConflict?: boolean;
  outro?: string | (() => string | Promise<string>);
  paths?: OptionsPaths;
  plugins?: OutputPlugin[];
  preferConst?: boolean;
  preserveModules?: boolean;
  preserveModulesRoot?: string;
  sourcemap?: boolean | "inline" | "hidden";
  sourcemapExcludeSources?: boolean;
  sourcemapFile?: string;
  sourcemapPathTransform?: SourcemapPathTransformOption;
  strict?: boolean;
  systemNullSetters?: boolean;
}
// ...
export interface SerializedTimings {
  [label: string]: [number, number, number];
}

export interface PreRenderedAsset {
  name: string | undefined;
  source: string | Uint8Array;
  type: "asset";
}

export interface OutputAsset extends PreRenderedAsset {
  fileName: string;
  /** @deprecated Accessing "isAsset" on files in the bundle is deprecated, please use "type === \'asset\'" instead */
  isAsset: true;
}

export interface RenderedModule {
  originalLength: number;
  removedExports: string[];
  renderedExports: string[];
  renderedLength: number;
}

export interface PreRenderedChunk {
  exports: string[];
  facadeModuleId: string | null;
  isDynamicEntry: boolean;
  isEntry: boolean;
  isImplicitEntry: boolean;
  modules: {
    [id: string]: RenderedModule;
  };
  name: string;
  type: "chunk";
}

export interface RenderedChunk extends PreRenderedChunk {
  code?: string;
  dynamicImports: string[];
  fileName: string;
  implicitlyLoadedBefore: string[];
  importedBindings: {
    [imported: string]: string[];
  };
  imports: string[];
  map?: SourceMap;
  referencedFiles: string[];
}

export interface OutputChunk extends RenderedChunk {
  code: string;
}

export interface SerializablePluginCache {
  [key: string]: [number, any];
}

export interface RollupCache {
  modules: ModuleJSON[];
  plugins?: Record<string, SerializablePluginCache>;
}

export interface RollupOutput {
  output: [OutputChunk, ...(OutputChunk | OutputAsset)[]];
}

export interface RollupBuild {
  cache: RollupCache | undefined;
  generate: (outputOptions: OutputOptions) => Promise<RollupOutput>;
  getTimings?: () => SerializedTimings;
  watchFiles: string[];
  write: (options: OutputOptions) => Promise<RollupOutput>;
}